requirements.txt
apex==0.9.10dev
deepspeed==0.6.0
detectron2.egg==info
fitz==0.0.1.dev2
fvcore==0.1.5.post20220212
jsonlines==3.0.0
numpy==1.21.5
omegaconf==2.1.1
opencv_python==4.5.5.62
Pillow==9.2.0
pycocoevalcap==1.2
pycocotools==2.0.4
PyMuPDF==1.19.6
pyparsing==3.0.7
ray==1.13.0
scikit_learn==1.1.1
setuptools==58.0.4
torch==1.10.2
torchvision==0.11.3
tqdm==4.63.0
transformers==4.17.0
